style(model): move to folder
